[NEW] route(n,m,...,si,di,...) primitive.
The route primitive allows to define a routing circuit with n inputs and m outputs. The internal connections are described by a list of pair of numbers (si,di) where si is the input number (numbered from 1) and di is the output number (numbered from 1). The value 0 as a special meaning and indicates the absence of input or output. Here are some examples:
- route(1,1,1,1) is equivalent to _
- route(1,1,1,0) is equivalent to !
- route(2,2,1,2,2,1) is equivalent to (_,_<:!,_,-,!)